video-analysis-worker/requirements.txt
psycopg[binary]>=3.2,<4
pydantic>=2.11,<3
python-dotenv>=1.0,<2
PyYAML>=6.0,<7
requests>=2.32,<3
patchright>=1.52,<2
